+++ src/oogeso/dto/devices.py
@@ -173,3 +173,13 @@ class DeviceWellGasLiftData(DeviceData):
     injection_pressure: float = None  # 20 # MPa
     separator_pressure: float = None  # 2 # MPa
     model: ModelType = ModelType.WELL_GAS_LIFT
+
+
+class DeviceCarbonCaptureData(DeviceData):
+    carbon_capture_rate: float = None  # ccr
+    exhaust_gas_recirculation: float = None  # egr
+    compressor_pressure_in: float = None
+    compressor_pressure_out: float = None
+    compressor_eta: float = None
+    compressor_temp_in: float = None
+    model: ModelType = ModelType.CARBON_CAPTURE