+++ src/oogeso/dto/mapper.py
@@ -49,6 +49,7 @@ def get_device_from_model_name(model_name: str) -> Callable:
         "storagehydrogen": devices.StorageHydrogen,
         "wellgaslift": devices.WellGasLift,
         "wellproduction": devices.WellProduction,
+        "carboncapture": devices.CarbonCapture,
     }
     if model_name in map_device_name_to_class:
         return map_device_name_to_class[model_name]
@@ -86,6 +87,7 @@ def get_device_data_class_from_str(model_name: str) -> Callable:
         "storagehydrogen": dto.DeviceStorageHydrogenData,
         "wellgaslift": dto.DeviceWellGasLiftData,
         "wellproduction": dto.DeviceWellProductionData,
+        "carboncapture": dto.DeviceCarbonCaptureData,
     }
     if model_name in map_device_name_to_class:
         return map_device_name_to_class[model_name]