+++ docs/changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## Unpublished
+2024-04-02
+- Add carbon capture and storage (CCS) devices
+
 ## v1.1.0
 2023-08-16
 - Add carbon network to keep track of emissions more easily (and allow CCS in the future)